Switch to a go-release-action for building binaries
For now, this uses a fork off the action on GitHub. It adds support
for leaving the binaries uncompressed. A PR has been opened on the
upstream, and we will switch to that once that lands (assumng it
does).

This patch also changes the trigger to "on release published", and
removes the "push to tags or branch ghactions".

Fixes #578

Signed-off-by: Brad P. Crochet <brad@redhat.com>